Move frontend to Frontend.hs
The previous place was an historical accident,
this is the right place for the driver.
